Move executor to GenericJob
